agent-project/3.ai_real_estate_agent/requirements.txt
streamlit
openai
agno
firecrawl-py==1.9.0
pydantic
streamlit
